重命名 OtherRefer 和 MyRefer
