+++ src/views/MyReferList.vue
@@ -1,9 +1,9 @@
 <template>
-    <div class="my-refer-list">
-        <div class="refers" v-loading="loading" element-loading-text="加载我的内推">
+    <div>
+        <div v-loading="loading" element-loading-text="加载我的内推">
             <ReferItem v-for="refer in refers" :key="refer.id" :refer="refer"/>
         </div>
-        <div class="pages">
+        <div>
             <el-pagination
                 v-show="totalPages !== 0"
                 :current-page.sync="page"
@@ -48,6 +48,4 @@
   }
 </script>
 
-<style scoped lang="scss">
-
-</style>
+<style scoped lang="scss"></style>
